# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionPlan": {
        "properties": {
          "actions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Action"
            },
            "title": "Actions",
            "type": "array"
          }
        },
        "title": "ActionPlan",
        "type": "object"
      },
...
      "BriefRequest": {
        "description": "Body for `/brief`, `/context`, `/verify`, `/actions/compose`.",
        "properties": {
          "intent": {
            "default": "prepare_decision_brief",
            "title": "Intent",
            "type": "string"
          },
          "user_id": {
            "default": "u_rm",
            "title": "User Id",
            "type": "string"
          }
        },
        "title": "BriefRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ConnectWork Command Agent \u2014 gateway",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/actions/compose": {
      "post": {
        "description": "Agent Actions; every action carries a previewable diff or a `blocked_reason`.",
        "operationId": "post_actions_compose_actions_compose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Actions Compose"
      }
    },
    "/actions/execute": {
      "post": {
        "description": "Execute only approved, non-blocked actions. The plan is recomposed server-side, so a\nblocked action is never executed even if its index is submitted as approved.",
        "operationId": "post_actions_execute_actions_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  },
                  "title": "Response Post Actions Execute Actions Execute Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Actions Execute"
      }
    },
    "/actions/loop": {
      "post": {
        "description": "Run the controlled work loop and return the dossier (assignments \u00b7 replies \u00b7 escalations \u00b7\nscheduled \u00b7 approved_indices \u00b7 audit \u00b7 closed, plus bundle/brief/plan).\n\nNOTE: `closed=True` means the loop CYCLE completed \u2014 not that every item is resolved. Derive\n'Open \u2014 escalation in flight' from `escalations` + any remaining blocked actions, not `closed`.",
        "operationId": "post_actions_loop_actions_loop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Actions Loop"
      }
    },
    "/actions/staged-remediation": {
      "post": {
        "description": "Validate one staged Decision Brief row remediation into one drawer card.\n\nThis endpoint is the anti-drift seam: the card comes from the row's remediation descriptor\npassed through the deterministic action composer, not a separately authored follow-up list.",
        "operationId": "post_actions_staged_remediation_actions_staged_remediation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StagedRemediationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Action"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Actions Staged Remediation"
      }
    },
    "/actions/staged-remediation/execute": {
      "post": {
        "description": "Execute exactly one staged Decision Brief row remediation by origin.",
        "operationId": "post_actions_staged_remediation_execute_actions_staged_remediation_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StagedRemediationExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StagedRemediationExecuteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Actions Staged Remediation Execute"
      }
    },
    "/api/actions": {
      "get": {
        "description": "The composed ActionPlan (typed diffs; blocked actions carry a `blocked_reason`).",
        "operationId": "api_actions_api_actions_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "u_rm",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "intent",
            "required": false,
            "schema": {
              "default": "prepare_decision_brief",
              "title": "Intent",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Actions Api Actions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Actions"
      }
    },
    "/api/brief": {
      "get": {
        "operationId": "api_brief_api_brief_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "u_rm",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "intent",
            "required": false,
            "schema": {
              "default": "prepare_decision_brief",
              "title": "Intent",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Brief Api Brief Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Brief"
      }
    },
...
    "/api/lifecycle": {
      "get": {
        "operationId": "api_lifecycle_api_lifecycle_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "u_rm",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "intent",
            "required": false,
            "schema": {
              "default": "prepare_decision_brief",
              "title": "Intent",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Lifecycle"
      }
    },
    "/api/lifecycle/events": {
      "post": {
        "operationId": "api_lifecycle_events_api_lifecycle_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifecycleEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Lifecycle Events"
      }
    },
...
    "/api/loop": {
      "get": {
        "description": "Convenience GET: same dossier as `POST /actions/loop` with the loop's default approval\n(`approve_nonblocked`); `closed` still means cycle-completed, not fully-resolved.",
        "operationId": "api_loop_api_loop_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "u_rm",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "intent",
            "required": false,
            "schema": {
              "default": "prepare_decision_brief",
              "title": "Intent",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Loop"
      }
    },
    "/api/meeting": {
      "get": {
        "operationId": "api_meeting_api_meeting_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "u_rm",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "intent",
            "required": false,
            "schema": {
              "default": "prepare_decision_brief",
              "title": "Intent",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Meeting Api Meeting Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Meeting"
      }
    },
...
    "/brief": {
      "post": {
        "description": "The DecisionBrief. `policy_gates` is the deterministic decision passed through untouched \u2014\nthe LLM layer never marks a brief approval-ready (the Acme case stays approval_ready=False).",
        "operationId": "post_brief_brief_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionBrief"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Brief"
      }
    },
    "/chat": {
      "post": {
        "description": "Answer a question about the meeting/decision using ONLY the permission-filtered bundle.\n\nThe LLM may draft prose, but the wrapper owns governance deterministically: restricted sources\nare refused (never revealed), the deterministic gate can't be overridden, citations are\nvalidated against the bundle, and missing evidence is surfaced honestly. Offline/deterministic\nby default (no API key); history is untrusted context, never evidence.",
        "operationId": "post_chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Chat"
      }
    },
    "/context": {
      "post": {
        "description": "The permission-filtered ContextBundle (debugging aid).",
        "operationId": "post_context_context_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Context"
      }
    },
...
  }
}